fix: remove lazy imports in models
