+++ sigstore/_internal/rekor/__init__.py
@@ -30,8 +30,7 @@ from sigstore._utils import base64_encode_pem_cert
 from sigstore.dsse import Envelope
 from sigstore.hashes import Hashed
 
-if typing.TYPE_CHECKING:
-    from sigstore.models import TransparencyLogEntry
+from sigstore._internal.rekor.entry import TransparencyLogEntry
 
 __all__ = [
     "_hashedrekord_from_parts",
