+++ sigstore/_internal/rekor/checkpoint.py
@@ -29,9 +29,8 @@ from pydantic import BaseModel, Field, StrictStr
 from sigstore._utils import KeyID
 from sigstore.errors import VerificationError
 
-if typing.TYPE_CHECKING:
-    from sigstore._internal.trust import RekorKeyring
-    from sigstore.models import TransparencyLogEntry
+from sigstore._internal.rekor.entry import TransparencyLogEntry
+from sigstore._internal.trust import RekorKeyring
 
 
 @dataclass(frozen=True)
